Fix missing columns from unused columns list
I checked the commit history going back to when this was still
a part of tipresias, and these just magically disappeared from
one commit to the next. I suspect a dodgy merge conflict
resolution since the disappearance happens around a PR merge
reversion, which means I was probably trying to fix stuff in a
panic.