Fix date parsing to avoid changing dates
This was causing the merging of player & match data to silently
fail, which resulted in bad data reaching the end of the pipeline.
I fixed the date-changing timezone conversions by adding default
times to match dates, and I changed the merge column to a
date-only column to be on the safe side.