Retrain & save models AGAIN
It turns out that someone introduced a breaking change in
scikit-learn 0.21.3 where any use of ColumnTransformer in
a pickled model fitted before this version will break, because
self._n_features is defined in fit_transform, then used in
transform. This is why pylint tells you not to do stuff like
this.