Fix player pipeline bugs
For some reason, the date parser stopped being able to handle
time strings with the format 'hhmm', so I manually separate them
with a ':'. Also, the AFL roster page is showing team lists from
last week, which is breaking the player data join, so I added a
filter for future-ish matches only.