Sort data columns in MLData
We were having issues with columns being sorted differently
when training vs predicting, and the cause was the node
combine_data when it was combining horizontally, because it
sorts data frames by length before joining. This is an issue in
prod because we truncate prod data sets to reduce memory usage,
which changes the sort-by-length order.

The validity of model inputs shouldn't depend on the vagaries
of kedro pipelines anyway, so it makes sense for MLData to be
responsible for model-specific data cleaning like this.