Return predictions in response rather than posting them
I added the POST when trying to implement augury's make_predictions
as a background process, but now that I've fixed it by extending
the timeout, it doesn't make any sense trying to post to an app
that we already have an open HTTP connection with. In fact, it
just raises an error.