Refactor rolling/expanding calculations into a method
We do this calculation of rolling values + filling of expanding
values in a few places, so it makes sense to encapsulate it in a
method. I had to change `group_keys` to True to make behaviour
consistent between single- and multi-indices.

The latest version of pandas must have changed how rolling vs
expanding indices are created, because this worked fine before
without the extra index munging.