Include fixture from up to a week ago
Setting the filter at the beginning of the week has caused errors,
because rounds aren't lasting full weeks at the moment, so there
are missing matches for the current round in the fixture data set.

We leave the filter for past matches at the beginning of the
current week to make sure we don't accidentally replace valid
match data with older fixture data. Deduplication in the
combine_data node should handle any overlap.