Fix flaky test for date filter
Two things were contributing to the flakiness. One, the filter
used Melbourne time for the timezone, but imported data gets
converted to UTC. Two, the end date filter that was used to check
the results was for the beginning of the day instead of the end.
These two oversights meant that match dates toward the edge of
the filter range could be erroneously included/excluded from the
filtered data set.