Save params for existing models separate from estimator classes
Now that new versions of the estimators will use the existing
estimator classes, we'll want to preserve the params for previous
versions to serve as benchmarks while we update default params
to reflect the new versions. We can't rely on the pickled
estimators, because we periodically update dependencies that
require re-saving picked estimators.