Downgrade statsmodels to 0.11.x
This is the last version we used before upgrading to 0.12.x,
which has a bunch of API changes that break our models. I kind
of want to replace the ARIMA model anyway, because the params
are way to sensitive, so I don't feel like figuring out this
new version with different params. Also, it would break the
2020 version of Tipresias, which is also annoying.