Delete full_data & save models sequentially
Running this script was crashing on occasiona, and I suspect it
was due to running out of memory, so better to free some up.
It kept happening after deleting the unused data, so I started
saving the models sequentially rather than in parallel. A bit
slower, but not as slow as having to run the script repeatedly
due to random crashes.