Reduce memory usage of integration tests
I was thinking that docker-compose was hanging, but it's
looking like the integration tests hang specifically once they
reach the player data tests, which use more memory than the
others, because it's a bigger data set.

GitHub actions hanging due to running out of memory would make
sense in this context. Especially since the way we define date
ranges for fetching data maximise data retrieved right before a
new season starts. To fix this, I've reduced the date range to
only fetch about a month of data when testing the player data
import.