Change cache key from branch name to commit SHA
The tutorial that I followed used the SHA, but I figured tying
the cache to the branch made more sense to avoid unnecessary
rebuilds. Now the post-docker/caching actions are hanging, and
an issue suggested that it has something to do with re-using
cache keys, doing stuff in parallel, etc. So, I'm changing the
key back to the original tutorial hoping that fixes it.