Change cache key used for Docker image
Hashing files makes for a better hash key than the commit SHA,
because if the dependencies don't change, we don't need to change
much from the stored image. Also, following a workaround from
a forum post, I'm adding a secret variable to reset the cache
manually when it gets stuck in a bad state like this.