Fix how Kedro session is loaded
We create the session at the various entrypoints to the app,
then get the current one from any of the internal modules/classes.
This required removing/rewriting all the code that uses the
context object directly, including removing a test that had
been skipped for long enough to render it meaningless.