Fix logging in production
There are some loggers in 'base' that we don't overwrite in
'production', so we get errors where they try to write to
non-existent log files. There isn't anything in the documentation,
but I'm hoping disable_existing_loggers will take care of it.
If not, I'll have to reorganise the logging config a bit.