Bump setup-gcloud back up to 0.2.1
I figured out the source of the problem with the missing keyfile:
the setup-gcloud action probably changed the order of its
processes, such that it now updates the
GOOGLE_APPLICATION_CREDENTIALS env var before exporting the
credentials to a file, meaning that my definition of the env var
never gets used. To fix this, we now start with a custom env var
that won't get overwritten, so the keyfile gets written to where
we expect to find it.