Remove version guards from responses package
Having version guards on packages adds extra admin work, and
there's not much benefit when the package is used in tests,
because a failing test will tell you that there's a breaking
change. It does make debugging a bit harder, since you won't
know which package broke the test right away, but that's often
the case even when you're careful with versions.