Tune params for new season
I haven't had time to actually work on the margin model, so I'm
reusing last-year's model with re-tuned params for the latest
seasons. The best score from CV is a bit lower than last year's,
but CV now includes the 2020 season, which, for obvious reasons,
was damn difficult to predict.

I removed the win-probability model, because it was buggy as hell
and has had two seasons of terrible predictions. So, my options
were to build something new from scratch or scrap it till I have
time to figure out a model that actually kind of works, which I
don't have time for this year, so maybe try again next year.